Added function to track particles in ring trap
The function can track particles moving around the ring trap using a
predictor that applies a bias when the particle is in the ring trap.
When the particle is outside the ring trap the bias is smaller and
eventually goes to zero at far distances from the average mean r of the
ring trap.